+++ web/scripts/dev.js
@@ -1,15 +1,18 @@
 const path = require("path")
 const { performance } = require("perf_hooks")
+const readline = require("readline")
 const formatMS = require("pretty-ms")
 const {
+  chalk,
   error,
   info,
   infoline,
   section,
   linebreak,
   question,
+  cmd,
   shell,
-  cmd
+  separator
 } = require("./pretty-logs")
 
 const DIR = path.resolve(__dirname, "../")
@@ -20,90 +23,167 @@ const isSudo = args.includes("sudo")
 
 const { createServer } = require("vite")
 
-;(async () => {
-  const start = performance.now()
+// -- MAIN
 
+let shuttingDown = false
+
+async function startup() {
   const doBuild = await question("Build containers? This can take a while. [y/N] -> ")
-  linebreak()
 
   if (doBuild.trim().toLowerCase() === "y") {
-    section("BUILD")
-
-    if (isSudo) cmd("pnpm build-sudo:local")
-    else cmd("pnpm run build:local")
-
-    infoline(
-      "Finished building containers.",
-      `Took: ${formatMS(performance.now() - start)}`
-    )
+    await buildContainers()
+    linebreak()
   } else {
     infoline("Skipping build.")
   }
 
-  section("DEV")
+  section("STARTUP", false, true)
+
+  await startVite()
+
   linebreak()
 
-  const legacyShell = shell("node scripts/build-legacy.js dev")
+  await startContainers()
 
-  const server = await createServer({ base: "/files--dev/" })
-  await server.listen()
+  infoline(
+    "Development servers started.",
+    "Don't forget that you can also start a Mockoon server to mock unimplemented API paths."
+  )
+
+  section("RUNNING", false, true)
+
+  if (process.stdin.isTTY) {
+    readline.emitKeypressEvents(process.stdin)
+    process.stdin.setRawMode(true)
+  }
+
+  logContainers()
+}
+
+async function shutdown() {
+  if (shuttingDown) return
+  shuttingDown = true
+
+  section("SHUTDOWN", true)
+
+  await stopVite()
+
+  linebreak()
+
+  await stopContainers()
+
+  infoline("Shutdown finished.")
+}
+
+// -- VITE LIFECYCLE
+
+let vite
+let viteStopping = false
+
+async function startVite() {
+  if (vite) return
+
+  vite = await createServer({ base: "/files--dev/" })
+  await vite.listen()
 
   // check for config not resolving correctly
   // this property should be defined
-  if (server.config?.server?.fs?.strict !== false) {
+  if (vite.config?.server?.fs?.strict !== false) {
     error("Vite didn't find its configuration file!")
-    process.exit(1)
+    await vite.close()
+    return process.exit()
   }
 
+  console.log(`${chalk.green("Vite started")} (${chalk.blueBright(vite.config.mode)})`)
+  vite.printUrls()
+}
+
+async function stopVite() {
+  if (!vite || viteStopping) return
+
+  vite = null
+  viteStopping = true
+
+  // prettier-ignore
+  try { await vite.close() } catch {}
+
+  info("Vite stopped.")
+}
+
+// -- CONTAINER LIFECYCLE
+
+let containersStarted = false
+let containersStopping = false
+let containerLogger = null
+
+async function buildContainers() {
+  section("BUILD", true)
+
+  const start = performance.now()
+
+  if (isSudo) cmd("pnpm -s build-sudo:local")
+  else cmd("pnpm -s build:local")
+
   infoline(
-    "Development servers (Vite, Legacy) started.",
-    "Don't forget that you can start a Mockoon server to mock unimplemented API paths."
+    "Finished building containers.",
+    `Took: ${formatMS(performance.now() - start)}`
   )
 
-  section("COMPOSE")
+  cmd("node scripts/build-legacy.js build")
+}
 
-  const composeShell = isSudo ? shell("pnpm compose-sudo up") : shell("pnpm compose up")
+async function startContainers() {
+  if (containersStarted) return
+  containersStarted = true
+  info("Starting containers...")
+  separator()
+  compose("up -d")
+}
 
-  // setup cleanup handlers
-  // we do this _past_ starting everything so that an early abort doesn't
-  // call these handlers and cause carnage in your terminal
+async function stopContainers() {
+  if (containersStopping) return
+  containersStopping = true
+  info("Stopping containers...")
+  separator()
 
-  process.once("cleanup", () => {
-    infoline("Cleaning up...")
+  if (containerLogger) {
+    containerLogger.kill()
+    containerLogger = null
+  }
 
-    if (process.platform === "win32") {
-      info(
-        "Since you're on Windows, the process might exit before the containers could stop.",
-        "Make sure to let the containers finish exiting prior to running 'pnpm dev' again.",
-        "If you want to always exit the program without dealing with any prompts,",
-        "make sure to HOLD Ctrl+C, rather than tapping it."
-      )
-      linebreak()
+  compose("stop")
+}
+
+async function logContainers() {
+  if (containerLogger) return
+  containerLogger = compose("logs -f --tail 10", true)
+}
+
+function compose(args, asShell) {
+  const str = `pnpm -s compose${isSudo ? "-sudo" : ""} -- ${args}`
+  return asShell ? shell(str) : cmd(str)
+}
+
+// -- RUN EVERYTHING
+
+// manually handle Ctrl+C
+process.stdin.on("keypress", async (str, key) => {
+  if (key.name === "c" && key.ctrl) {
+    // spamming Ctrl+C again will turn off raw mode,
+    // and will try to exit the process again
+    if (shuttingDown) {
+      process.stdin.setRawMode(false)
+    } else {
+      await shutdown()
     }
+    process.exit(0)
+  }
+})
 
-    server.close()
-    legacyShell.kill()
-    composeShell.kill()
-  })
+// make very sure we stop the damn server
+process.once("beforeExit", shutdown)
+process.once("SIGINT", shutdown)
+process.once("SIGTERM", shutdown)
+process.once("SIGHUP", shutdown)
 
-  // do app specific cleaning before exiting
-  process.once("beforeExit", () => {
-    process.emit("cleanup")
-  })
-
-  // catch closing terminal on Windows
-  process.once("SIGHUP", () => {
-    process.emit("cleanup")
-  })
-
-  // catch ctrl+c event
-  process.once("SIGINT", () => {
-    process.emit("cleanup")
-  })
-
-  // catch uncaught exceptions
-  process.once("uncaughtException", err => {
-    process.emit("cleanup")
-    throw new Error(err)
-  })
-})()
+startup()