+++ web/modules/ftml-wasm-worker/CHANGELOG.md
@@ -1,6 +1,24 @@
 # @wikijump/ftml-wasm-worker
 
+## 0.2.0
+
+### Minor Changes
+
+- Overdue versioning bump
+
+### Patch Changes
+
+- 5ee0dde63: Change how the FTML WASM is imported by default
+- Updated dependencies [5ee0dde63]
+- Updated dependencies
+- Updated dependencies [5ee0dde63]
+- Updated dependencies [5ee0dde63]
+  - @wikijump/util@0.2.0
+  - @wikijump/ftml-wasm@1.7.0
+  - @wikijump/threads-worker-module@0.1.3
+
 ## 0.1.5
+
 ### Patch Changes
 
 - Updated dependencies [1ebea8151]